tests: adiciona testes para funções que verifica permissoes, corrige testes das avaliacoes e reorganiza as pastas de testes
