--- teste_evento
@@ -1,1 +1,0 @@
-Avaliação criada com sucesso!
