+++ tests/testes_unitarios/estrutura_de_provas/avaliacao/test_agregado.py
