+++ tests/testes_unitarios/estrutura_de_provas/exercicios/test_agregado.py
